+++ cmd/provenanced/cmd/testnet_test.go
@@ -11,10 +11,11 @@ import (
 	banktypes "github.com/cosmos/cosmos-sdk/x/bank/types"
 	genutiltest "github.com/cosmos/cosmos-sdk/x/genutil/client/testutil"
 	genutiltypes "github.com/cosmos/cosmos-sdk/x/genutil/types"
-	"github.com/provenance-io/provenance/app"
 	"github.com/spf13/viper"
 	"github.com/stretchr/testify/require"
 	"github.com/tendermint/tendermint/libs/log"
+
+	"github.com/provenance-io/provenance/app"
 )
 
 func Test_TestnetCmd(t *testing.T) {