+++ x/attribute/client/cli/query.go
@@ -39,16 +39,10 @@ func GetQueryCmd() *cobra.Command {
 // GetAttributeParamsCmd returns the command handler for name parameter querying.
 func GetAttributeParamsCmd() *cobra.Command {
 	cmd := &cobra.Command{
-		Use:   "params",
-		Short: "Query the current name parameters",
-		Args:  cobra.NoArgs,
-		Long: strings.TrimSpace(
-			fmt.Sprintf(`Query the current attribute module parameters:
-
-$ %s query attribute params
-`,
-				version.AppName,
-			)),
+		Use:     "params",
+		Short:   "Query the current name parameters",
+		Args:    cobra.NoArgs,
+		Example: fmt.Sprintf(`$ %s query attribute params`, version.AppName),
 		RunE: func(cmd *cobra.Command, args []string) error {
 			clientCtx, err := client.GetClientQueryContext(cmd)
 			if err != nil {
@@ -75,14 +69,12 @@ func GetAccountAttributeCmd() *cobra.Command {
 	cmd := &cobra.Command{
 		Use:   "get [address] [name]",
 		Short: "Get account attributes by name",
-		Long: strings.TrimSpace(
-			fmt.Sprintf(`Query for all attributes on account with a given name:
-
-Example:
-$ %s query attribute get pb1skjwj5whet0lpe65qaq4rpq03hjxlwd9nf39lk attrib.name
-$ %s query attribute get pb1skjwj5whet0lpe65qaq4rpq03hjxlwd9nf39lk attrib.name --page=2 --limit=100
-`,
-				version.AppName, version.AppName,
+		Example: strings.TrimSpace(
+			fmt.Sprintf(`
+				$ %[1]s query attribute get pb1skjwj5whet0lpe65qaq4rpq03hjxlwd9nf39lk attrib.name
+				$ %[1]s query attribute get pb1skjwj5whet0lpe65qaq4rpq03hjxlwd9nf39lk attrib.name --page=2 --limit=100
+				`,
+				version.AppName,
 			)),
 		Args: cobra.ExactArgs(2),
 		RunE: func(cmd *cobra.Command, args []string) error {
@@ -123,14 +115,12 @@ func ListAccountAttributesCmd() *cobra.Command {
 	cmd := &cobra.Command{
 		Use:   "list [address]",
 		Short: "Get all account attributes",
-		Long: strings.TrimSpace(
-			fmt.Sprintf(`Query for all attributes on account with a given name:
-
-Example:
-$ %s query attribute list pb1skjwj5whet0lpe65qaq4rpq03hjxlwd9nf39lk
-$ %s query attribute list pb1skjwj5whet0lpe65qaq4rpq03hjxlwd9nf39lk --page=2 --limit=100
-`,
-				version.AppName, version.AppName,
+		Example: strings.TrimSpace(
+			fmt.Sprintf(`
+				$ %[1]s query attribute list pb1skjwj5whet0lpe65qaq4rpq03hjxlwd9nf39lk
+				$ %[1]s query attribute list pb1skjwj5whet0lpe65qaq4rpq03hjxlwd9nf39lk --page=2 --limit=100
+				`,
+				version.AppName,
 			)),
 		Args: cobra.ExactArgs(1),
 		RunE: func(cmd *cobra.Command, args []string) error {
@@ -169,14 +159,12 @@ func ScanAccountAttributesCmd() *cobra.Command {
 	cmd := &cobra.Command{
 		Use:   "scan [address] [suffix]",
 		Short: "Scan account attributes by name suffix",
-		Long: strings.TrimSpace(
-			fmt.Sprintf(`Query for all attributes on account with a given name:
-
-Example:
-$ %s query attribute scan pb1skjwj5whet0lpe65qaq4rpq03hjxlwd9nf39lk name.suffix
-$ %s query attribute scan pb1skjwj5whet0lpe65qaq4rpq03hjxlwd9nf39lk name.suffix --page=2 --limit=100
-`,
-				version.AppName, version.AppName,
+		Example: strings.TrimSpace(
+			fmt.Sprintf(`
+				$ %[1]s query attribute scan pb1skjwj5whet0lpe65qaq4rpq03hjxlwd9nf39lk name.suffix
+				$ %[1]s query attribute scan pb1skjwj5whet0lpe65qaq4rpq03hjxlwd9nf39lk name.suffix --page=2 --limit=100
+				`,
+				version.AppName,
 			)),
 		Args: cobra.ExactArgs(2),
 		RunE: func(cmd *cobra.Command, args []string) error {