+++ x/marker/keeper/migrations.go
@@ -2,6 +2,7 @@ package keeper
 
 import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
+
 	v042 "github.com/provenance-io/provenance/x/marker/legacy/v042"
 )
 
@@ -18,8 +19,11 @@ func NewMigrator(keeper Keeper) Migrator {
 // Migrate1to2 migrates from version 1 to 2.
 func (m *Migrator) Migrate1to2(ctx sdk.Context) error {
 	ctx.Logger().Info("Migrating Marker Module from Version 1 to 2")
-	if err := v042.MigrateMarkerAddressKeys(ctx, m.keeper.storeKey); err != nil {
+	err := v042.MigrateMarkerAddressKeys(ctx, m.keeper.storeKey)
+	if err != nil {
 		return err
 	}
-	return v042.MigrateMarkerPermissions(ctx, m.keeper)
+	err = v042.MigrateMarkerPermissions(ctx, m.keeper)
+	ctx.Logger().Info("Finished Migrating Marker Module from Version 1 to 2")
+	return err
 }