+++ x/name/keeper/migrations.go
@@ -2,6 +2,7 @@ package keeper
 
 import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
+
 	v042 "github.com/provenance-io/provenance/x/name/legacy/v042"
 )
 
@@ -17,6 +18,8 @@ func NewMigrator(keeper Keeper) Migrator {
 
 // Migrate1to2 migrates from version 1 to 2.
 func (m *Migrator) Migrate1to2(ctx sdk.Context) error {
-	ctx.Logger().Info("Migrating Name Module from Version 1 to 2")
-	return v042.MigrateAddresses(ctx, m.keeper.storeKey)
+	ctx.Logger().Info("Migrating Name Module from Version 1 to 2 (1/1)")
+	err := v042.MigrateAddresses(ctx, m.keeper.storeKey)
+	ctx.Logger().Info("Finished Migrating Name Module from Version 1 to 2")
+	return err
 }